+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
             android:name=".MainActivity"
             android:configChanges="orientation|keyboardHidden|screenSize"
             android:label="@string/app_name"
-            android:theme="@style/FullscreenTheme">
+            android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -20,13 +20,9 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".Main2Activity"
-            android:label="@string/title_activity_main2"
+            android:name=".VSCpuActivity"
+            android:label="VS CPU"
             android:theme="@style/AppTheme.NoActionBar" />
-        <activity
-            android:name=".SinglePlayerPrototype"
-            android:label="@string/title_activity_single_player_prototype"
-            android:theme="@style/AppTheme.NoActionBar"></activity>
     </application>
 
 </manifest>
