+++ app/src/main/res/layout/activity_main.xml
@@ -1,12 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
+    android:launchMode= "singleTask"
     tools:context="app.simone.MainActivity"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
     <LinearLayout
-        android:id="@+id/fullscreen_content"
+        android:id="@+id/main_fullscreen_content"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
@@ -22,16 +23,16 @@
             <Button
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:background="#1ab855"
-                android:textColor="#fff"
+                android:background="@color/myGreen"
+                android:textColor="@color/myMainTextColor"
                 android:layout_weight="1"
                 android:text="Player vs CPU"
-                android:id="@+id/button1"/>
+                android:id="@+id/button_vs_cpu"/>
             <Button
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:background="#FF4500"
-                android:textColor="#fff"
+                android:background="@color/myRed"
+                android:textColor="@color/myMainTextColor"
                 android:layout_weight="1"
                 android:text="Player vs Player"
                 android:id="@+id/button2"/>
@@ -48,8 +49,8 @@
             <Button
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:background="#75d4f0"
-                android:textColor="#fff"
+                android:background="@color/myBlue"
+                android:textColor="@color/myMainTextColor"
                 android:layout_weight="1"
                 android:text="Highscore"
                 android:id="@+id/button3"/>
@@ -57,8 +58,8 @@
             <Button
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:background="#FFD700"
-                android:textColor="#fff"
+                android:background="@color/myYellow"
+                android:textColor="@color/myMainTextColor"
                 android:layout_weight="1"
                 android:text="Settings"
                 android:id="@+id/button4"/>
