+++ app/src/main/res/values/colors.xml
@@ -3,6 +3,11 @@
     <color name="colorPrimary">#3F51B5</color>
     <color name="colorPrimaryDark">#303F9F</color>
     <color name="colorAccent">#FF4081</color>
-
+    <color name="myGreen">#1ab855</color>
+    <color name="myDarkGreen">#13863d</color>
+    <color name="myRed">#FF4500</color>
+    <color name="myBlue">#75d4f0</color>
+    <color name="myYellow">#FFD700</color>
+    <color name="myMainTextColor">#fff</color>
     <color name="black_overlay">#66000000</color>
 </resources>
