+++ app/src/main/res/values/strings.xml
@@ -5,4 +5,5 @@
     <string name="dummy_content">DUMMY\nCONTENT</string>
     <string name="title_activity_main2">Main2Activity</string>
     <string name="title_activity_single_player_prototype">SinglePlayerPrototype</string>
+    <string name="vs_cpu">VS\nCPU</string>
 </resources>
